+++ test/helpers.js
@@ -1,0 +1,10 @@
+'use strict';
+
+exports.expectError = async resolver => {
+  try {
+    await resolver();
+  } catch (e) {
+    return;
+  }
+  throw new Error('missing expected error');
+};